Исправил Film и его вытекающие

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -45,7 +45,7 @@ public class FilmService {
     }
 
     public Optional<Film> findById(int id) {
-        return filmStorage.findById((long) id);
+        return filmStorage.findById((int) id);
     }
 
     public Film add(Film film) {
